Refactor Adapter parts into Manager
The Manager only instantiated an Adapter. Just bring the bits of Adapter up into Manager.
